test-agent/pyproject.toml
[project]
name = "openbox-langgraph-test-agent"
version = "1.0.0"
description = "Test LangGraph agent using the OpenBox LangGraph Python SDK"
requires-python = ">=3.11"
dependencies = [
    "openbox-langgraph-sdk>=0.1.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
openbox-langgraph-sdk = { path = "../", editable = true }
